feat: rollup for package bundling
