+++ src/handlers/index.ts
@@ -1,0 +1,2 @@
+export * from "./generate-erc20-permit";
+export * from "./generate-erc721-permit";
